refactor(dashboard): Further reduce duplication
